+++ flake.nix
@@ -5,9 +5,11 @@
     flake-utils.url = "github:numtide/flake-utils";
     flake-compat.url = "github:edolstra/flake-compat";
     flake-compat.flake = false;
+    npmlock2nix.url = "github:nix-community/npmlock2nix";
+    npmlock2nix.flake = false;
   };
 
-  outputs = { self, flake-utils, nixpkgs, ... }:
+  outputs = { self, flake-utils, npmlock2nix, nixpkgs, ... }:
     let
       attrs = nixpkgs.lib.importJSON ./package.json;
       inherit (attrs) name version;
@@ -16,11 +18,15 @@
     flake-utils.lib.eachDefaultSystem (system:
       let
         pkgs = import nixpkgs { inherit system; };
+        npm = pkgs.callPackage npmlock2nix { };
+        node_modules_attrs = {
+          buildInputs = [ pkgs.python3 pkgs.pkg-config pkgs.libsecret ];
+        };
       in
       {
         defaultPackage = self.packages.${system}.vsix;
 
-        packages.vsix = pkgs.mkYarnPackage {
+        packages.vsix = npm.build {
           src = ./.;
           name = vsix;
           buildPhase = ''
@@ -37,8 +43,9 @@
           distPhase = ":";
         };
 
-        devShell = pkgs.mkShell {
-          inputsFrom = [ self.defaultPackage.${system} ];
+        devShell = npm.shell {
+          src = ./.;
+          inherit node_modules_attrs;
         };
       });
 }