+++ src/test/suite/direnv.test.ts
@@ -1,8 +1,9 @@
-import * as assert from 'assert'
+import { strict as assert } from 'assert'
 import * as path from 'path'
 import * as sinon from 'sinon'
 import * as vscode from 'vscode'
 import { workspaceRoot } from '.'
+import * as config from '../../config'
 import * as direnv from '../../direnv'
 
 describe('direnv', () => {
@@ -17,20 +18,25 @@ describe('direnv', () => {
 	})
 
 	describe('in the test workspace', () => {
+		it('finds the direnv executable', async () => {
+			await direnv.test()
+		})
+
 		it('finds the .envrc file in the workspace root', async () => {
 			const path = await direnv.find()
-			assert.strict.equal(path, file)
+			assert.equal(path, file)
 		})
 
 		it('reuses the .envrc file in the workspace root', async () => {
 			const path = await direnv.create()
-			assert.strict.equal(path, file)
+			assert.equal(path, file)
 		})
 
 		it('dumps the allowed .envrc file', async () => {
+			delete process.env['VARIABLE']
 			await direnv.allow(file)
 			const data = await direnv.dump()
-			assert.strict.equal(data['VARIABLE'], 'value')
+			assert.equal(data['VARIABLE'], 'value')
 		})
 
 		it('fails to dump the blocked .envrc file', async () => {
@@ -39,8 +45,14 @@ describe('direnv', () => {
 			try {
 				await direnv.dump()
 			} catch ({ path }) {
-				assert.strict.equal(path, file)
+				assert.equal(path, file)
 			}
+		})
+
+		it('fails when the direnv executable is missing', async () => {
+			const missing = '/missing/executable'
+			sinon.replace(config.path.executable, 'get', () => missing)
+			await assert.rejects(() => direnv.test(), new direnv.CommandNotFoundError(missing))
 		})
 	})
 
@@ -49,17 +61,19 @@ describe('direnv', () => {
 		const subfile = path.join(subdir, '.envrc')
 
 		beforeEach(() => {
-			sinon.replaceGetter(vscode.workspace, 'workspaceFolders', () => [{ index: 0, name: 'subdir', uri: vscode.Uri.file(subdir) }])
+			sinon.replaceGetter(vscode.workspace, 'workspaceFolders', () => [
+				{ index: 0, name: 'subdir', uri: vscode.Uri.file(subdir) },
+			])
 		})
 
 		it('finds the .envrc file in the parent directory', async () => {
 			const path = await direnv.find()
-			assert.strict.equal(path, file)
+			assert.equal(path, file)
 		})
 
 		it('creates an .envrc file in the subdirectory', async () => {
 			const path = await direnv.create()
-			assert.strict.equal(path, subfile)
+			assert.equal(path, subfile)
 		})
 	})
 })